requirements.txt
python-binance==1.0.12
python-dotenv==0.17.1
PySocks==1.7.1